% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{f1_score}
\alias{f1_score}
\title{Calculate reconstruction errors. Function calculates separate reconstruction
errors for categorical and numerical variables}
\usage{
f1_score(y, yhat)
}
\arguments{
\item{Xhat}{Reconstructed dataset}

\item{X}{Ground truth dataset}
}
\value{
Column-wise reconstruction error, and the average reconstruction
error for categorical and numeric variables
}
\description{
Calculate reconstruction errors. Function calculates separate reconstruction
errors for categorical and numerical variables
}
