% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sparsify}
\alias{sparsify}
\title{Apply adaptive sparsity thresholds}
\usage{
sparsify(emap, A0, parallel = TRUE)
}
\arguments{
\item{emap}{Spectral embedding for the \code{rf} learned via \code{eigenmap}.}

\item{A0}{Adjacency matrix for test samples.}

\item{parallel}{Compute in parallel?}
}
\description{
This function caps the number of expected neighbors in a data-driven manner.
}
\keyword{internal}
