% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_knn.R
\name{decode_knn}
\alias{decode_knn}
\title{Decode RF Embeddings}
\usage{
decode_knn(rf, emap, z, x_tilde = NULL, k = 5, parallel = TRUE)
}
\arguments{
\item{rf}{Pre-trained random forest object of class \code{ranger}.}

\item{emap}{Spectral embedding learned via \code{eigenmap}.}

\item{z}{Matrix of embedded data to map back to the input space.}

\item{x_tilde}{Supplied training data, if none supplied then the RF is used
to generate synthetic training data according to the eForest scheme. Default
is NULL.}

\item{k}{Number of nearest neighbors to evaluate.}

\item{parallel}{Compute in parallel? Must register backend beforehand, e.g.
via \code{doParallel}.}
}
\value{
Decoded dataset.
}
\description{
Maps the low-dimensional KPCA embedding of a random forest back to the input
space via iterative k-nearest neighbors.
}
