% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{post_x}
\alias{post_x}
\title{Post-process data}
\usage{
post_x(x, meta, round = TRUE)
}
\arguments{
\item{x}{Input \code{data.frame}.}

\item{meta}{Metadata.}

\item{round}{Round continuous variables to their respective maximum precision
in the real data set?}
}
\value{
A \code{data.frame} which follows the structure and ordering of the input dataset.
}
\description{
This function prepares output data.
}
