% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prep_x}
\alias{prep_x}
\title{Preprocess input data}
\usage{
prep_x(x, to_numeric = NULL, to_factor = NULL, default = 5)
}
\arguments{
\item{x}{Input data.frame.}

\item{to_numeric}{List of variables to force as numeric.}

\item{to_factor}{List of variables to force as factor.}

\item{default}{Threshold to classify a variable as numeric (more than default
unique values) or factor (less or equal to unique values).}
}
\value{
Preprocessed data.frame.
}
\description{
This function prepares input data.
}
