% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prep_x}
\alias{prep_x}
\title{Preprocess input data}
\usage{
prep_x(x, to_numeric = NULL, to_factor = NULL, default = 5)
}
\arguments{
\item{x}{Input data.frame.}
}
\description{
This function prepares input data.
}
