% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{reconstruction_error}
\alias{reconstruction_error}
\title{Mixed-type Reconstruction Error}
\usage{
reconstruction_error(Xhat, X)
}
\arguments{
\item{Xhat}{Reconstructed dataset}

\item{X}{Ground truth dataset}
}
\value{
A list containing column-wise reconstruction accuracy, and the average
reconstruction error for categorical and numeric variables. Values lie
between 0-1, where 1 represents perfect reconstruction, and 0 represents no
reconstruction. Note that in the original paper, this is presented as the
distortion - this is because we do 1 - error in the plotting stage.
}
\description{
Mixed-type Reconstruction Error
}
\details{
In standard AEs, reconstruction error is generally estimated via \eqn{L_2}
loss. This is not sensible with a mix of continuous and categorical data, so
we devise a measure that evaluates distortion on continuous variables as
\eqn{1 - R^2}, and categorical variables as accuracy.
}
\examples{

arf <- adversarial_rf::arf(iris)
emap <- encode(arf, iris, k = 2)
z <- predict.encode(emap, arf, iris)
out <- decode_knn(arf, emap, z, k = 5)$x_hat
error <- reconstruction_error(out, X)

}
