% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_knn.R
\name{train_decoder}
\alias{train_decoder}
\title{Train RF Decoder}
\usage{
train_decoder(rf, emap, neighbors = NULL, null_value = NULL, parallel = TRUE)
}
\description{
This function rebuilds training data like eForest.
}
\keyword{internal}
