% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_knn.R
\name{train_decoder}
\alias{train_decoder}
\title{Decode RF Embeddings}
\usage{
train_decoder(rf, emap, neighbors = NULL, null_value = NULL, parallel = TRUE)
}
\arguments{
\item{rf}{Pre-trained random forest object of class \code{ranger}.}

\item{emap}{Spectral embedding learned via \code{eigenmap}.}

\item{parallel}{Compute in parallel? Must register backend beforehand, e.g.
via \code{doParallel}.}

\item{z}{Matrix of embedded data to map back to the input space.}

\item{k}{Number of nearest neighbors to evaluate.}
}
\value{
Decoded dataset.
}
\description{
Maps the low-dimensional KPCA embedding of a random forest back to the input
space via iterative k-nearest neighbors.
}
